requirements.txt
Flask
pymongo 
bcrypt
werkzeug
python-dotenv